drop the predicateArgument view and just use spanLabelView.
